Implemented Max Cut Weight GES Tabu algorithm Addded test graphs

+++ src/com/elvis/optimizationtask/algorithm/maxcut/MaxCut.java
@@ -9,7 +9,7 @@ package com.elvis.optimizationtask.algorithm.maxcut;
 public interface MaxCut {
     public long getTimeExec();
 
-    public long getMaxCut();
+    public float getMaxCut();
 
     public boolean[] getMask();
